HandyLink-backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt==5.2.2